Improve names, make model fitting more difficult.
I pushed the group responses a bit closer by adjusting the slope of the
drug group. Stan code is a bit cleaner, but more verbose than it might
need to be. Ideally, we'd precompute normal_lpdf in the transformed
parameters block, but we can't declare/modify the counter from there (n).